requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.21
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.2.0
spacy>=3.7.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0.tar.gz
reportlab>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.1.14
langchain-core>=0.1.32
langchain-community>=0.0.29
langchain-huggingface>=0.0.6
google-generativeai>=0.3.2
pydantic>=2.0
